Extract git data models
