+++ .vscode/settings.json
@@ -4,4 +4,5 @@
     "quickFix.biome": "explicit",
     "source.organizeImports.biome": "explicit"
   },
+  "typescript.tsdk": "node_modules\\typescript\\lib",
 }
