+++ src/components/layout/main-layout.tsx
@@ -10,7 +10,8 @@ import { useSettingsStore } from "@/settings/stores/settings-store";
 import { useBufferStore } from "@/stores/buffer-store";
 import { useUIState } from "@/stores/ui-state-store";
 import DiffViewer from "@/version-control/diff-viewer/views/diff-viewer";
-import { type GitHunk, stageHunk, unstageHunk } from "../../version-control/git/controllers/git";
+import { stageHunk, unstageHunk } from "@/version-control/git/controllers/git";
+import type { GitHunk } from "@/version-control/git/models/git-types";
 import AIChat from "../ai-chat/ai-chat";
 import BottomPane from "../bottom-pane";
 import CommandBar from "../command/components/command-bar";
