+++ src/version-control/git/controllers/git-diff-cache.ts
@@ -1,4 +1,4 @@
-import type { GitDiff } from "./git";
+import type { GitDiff } from "../models/git-types";
 
 interface CacheEntry {
   diff: GitDiff;
